refactor(sidebar): extract telemetry module
